Reduit les lignes de dates

Reduire les lignes d'une table selon la date de debut et la date de fin. Utiles pour analyser les differentes saisies d'absences ou les differents contrats agents qui se suivent, etc.

distinct_dates(
  data,
  groups,
  variable_date_debut,
  variable_date_fin,
  tronquer_debut = FALSE,
  tronquer_fin = FALSE,
  tolerance_weekend = TRUE,
  tolerance_nb_jours = 1
)

Arguments

data
une table contenant au moins une colonne date de debut et une colonne date de fin (voir autres arugments)

groups
le(s) (differents) groupe(s) dont il faut tenir compte variable d'identification (ex: Matricule, Statut)

variable_date_debut
une seule variable qui indique la date de debut (debut agent, debut absences, ...)

variable_date_fin
une seule variable qui indique la date de fin (fin agent, fin absences, ...)

tronquer_debut
Date entre guillemets "2018-01-01" dans le format "YYYY-MM-JJ" pour tronquer les dates de debut (utile pour le calcul du nombre de jours)

tronquer_fin
Date entre guillemets "2018-01-01" dans le format "YYYY-MM-JJ" pour tronquer les dates de fin (utile pour le calcul du nombre de jours)

tolerance_weekend
si TRUE, alors tolere une difference de jours deux lignes si cette difference est uniquement sur le weekend (date de fin un vendredi/samedi/dimanche et date de debut suivante est un lundi)

tolerance_nb_jours
default 1. Nombre de jours de tolerance entre la date de fin et la prochaine date de debut (par exemple 31 jours pour les contrats agents ou 1 jour pour les absences a la suite)

Value

une table reduite

Examples

distinct_dates(data, MATRICULE, variable_date_debut = `date debut agent`, variable_date_fin = `date fin agent`, tolerance_nb_jours = 31)

#> Le chargement a nécessité le package : dplyr

#> 
#> Attachement du package : ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

#> Le chargement a nécessité le package : magrittr

#> Le chargement a nécessité le package : lubridate

#> 
#> Attachement du package : ‘lubridate’

#> The following objects are masked from ‘package:dplyr’:
#> 
#>     intersect, setdiff, union

#> The following objects are masked from ‘package:base’:
#> 
#>     date, intersect, setdiff, union

#> Error in UseMethod("mutate_"): pas de méthode pour 'mutate_' applicable pour un objet de classe "function"

# Multiple variables de groupes
distinct_date(data, vars(MATRICULE, STATUT), date_debut, date_fin )

#> Error in distinct_date(data, vars(MATRICULE, STATUT), date_debut, date_fin): impossible de trouver la fonction "distinct_date"